+++ turbo/apps/platform/src/views/settings-page/connector-icons.tsx
@@ -27,8 +27,10 @@ import neonIcon from "./icons/neon.svg";
 import notionIcon from "./icons/notion.svg";
 import outlookCalendarIcon from "./icons/outlook-calendar.svg";
 import outlookMailIcon from "./icons/outlook-mail.svg";
+import posthogIcon from "./icons/posthog.svg";
 import redditIcon from "./icons/reddit.svg";
 import sentryIcon from "./icons/sentry.svg";
+import similarwebIcon from "./icons/similarweb.svg";
 import slackIcon from "./icons/slack.svg";
 import stravaIcon from "./icons/strava.svg";
 import stripeIcon from "./icons/stripe.svg";
@@ -67,8 +69,10 @@ const CONNECTOR_ICONS: Readonly<Record<ConnectorType, string>> = Object.freeze({
   notion: notionIcon,
   "outlook-calendar": outlookCalendarIcon,
   "outlook-mail": outlookMailIcon,
+  posthog: posthogIcon,
   reddit: redditIcon,
   sentry: sentryIcon,
+  similarweb: similarwebIcon,
   slack: slackIcon,
   strava: stravaIcon,
   stripe: stripeIcon,